requirements.txt
amqp==5.0.2
billiard==3.6.3.0
celery==5.0.3
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
Flask==1.1.2
flask-fs==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
kombu==5.0.2
MarkupSafe==1.1.1
Pillow==8.0.1
prompt-toolkit==3.0.8
pycparser==2.20
pymongo==3.11.2
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.4
redis==3.5.3
six==1.15.0
tzlocal==2.1
vine==5.0.0
wcwidth==0.2.5
Werkzeug==1.0.1
